Разбивает общий файл на модули
